% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answerlifeuniv.R
\name{ultimate_answer}
\alias{ultimate_answer}
\title{The Answer to the Ultimate Question of Life, the Universe, and Everything}
\usage{
ultimate_answer(x)
}
\arguments{
\item{x}{Any input, really}
}
\value{
The answer to the ultimate question, which is 42.
}
\description{
This is very self explanatory
}
\examples{
ultimate_answer(999999)
}
